+++ pkg/network/anonymity/queue/settings.go
@@ -14,6 +14,7 @@ type SSettings sSettings
 type sSettings struct {
 	FMessageConstructSettings net_message.IConstructSettings
 	FNetworkMask              uint32
+	FConsumersCap             uint64
 	FPoolCapacity             [2]uint64
 	FQueuePeriod              time.Duration
 }
@@ -22,6 +23,7 @@ func NewSettings(pSett *SSettings) ISettings {
 	return (&sSettings{
 		FMessageConstructSettings: pSett.FMessageConstructSettings,
 		FNetworkMask:              pSett.FNetworkMask,
+		FConsumersCap:             pSett.FConsumersCap,
 		FPoolCapacity:             pSett.FPoolCapacity,
 		FQueuePeriod:              pSett.FQueuePeriod,
 	}).mustNotNull()
@@ -36,6 +38,9 @@ func (p *sSettings) mustNotNull() ISettings {
 	}
 	if p.FPoolCapacity[0] == 0 || p.FPoolCapacity[1] == 0 {
 		panic(`p.FPoolCapacity[0] == 0 || p.FPoolCapacity[1] == 0`)
+	}
+	if p.FConsumersCap == 0 {
+		panic(`p.FConsumersLen == 0`)
 	}
 	// p.FParallel, p.FNetworkMask, p.FWorkSizeBits, p.FRandQueuePeriod, p.FLimitVoidSizeBytes can be = 0
 	return p
@@ -55,4 +60,8 @@ func (p *sSettings) GetPoolCapacity() [2]uint64 {
 
 func (p *sSettings) GetQueuePeriod() time.Duration {
 	return p.FQueuePeriod
+}
+
+func (p *sSettings) GetConsumersCap() uint64 {
+	return p.FConsumersCap
 }